Move currentAPI to its own file apiEndpoints
